Move admin testcase handlers to handler package
